Moved IsOpaque into camera service interface
